refactor transaction components
